pyproject.toml
[tool.poetry]
name = "kanbanize"
version = "0.1.0"
description = ""
authors = ["Mateusz Barańczyk <mateusz.baranczyk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
fastapi = "^0.110.0"
gunicorn = "^21.2.0"
uvicorn = "^0.28.0"
httpx = "^0.27.0"
google-cloud-firestore = "^2.15.0"
mock-firestore = "^0.11.0"
pika = "^1.3.2"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^3.3.1"

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79
target_version = ['py311']

[tool.flake8]
max-line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"